Add files via upload
Revised scripts used to process final version of the paper.
Note: A unit pressure for the water vapour correction was discovered during peer review which was biasing pCO2 values low. This has been fixed in these scripts.
Note: All scripts have been reformatted to improve readability. All data analysis in plot_Bergmann_underway_data.m is linked to the sections of the paper it relates to.